Finished input logic rewrite.
- Added InputContext, a class which allows for defining sets of inputs which can easily be added and removed, and also allow for blocking lower priority contexts.
- Added an Editor context, containing all editor actions.
- Added CameraComponent2D

+++ include/puffin/platform/raylib/input/inputsubsystemrl.h
@@ -34,8 +34,8 @@ namespace puffin::input
 		int GetMouseY() override;
 		Vector2i GetMousePosition() override;
 
-		double GetMouseDeltaX() override;
-		double GetMouseDeltaY() override;
+		float GetMouseDeltaX() override;
+		float GetMouseDeltaY() override;
 		Vector2f GetMouseDelta() override;
 
 	protected:
